double => float (byte issue)

+++ controlling/src/main/java/me/autobot/lib/serial/SensorHubI2CConnection.java
@@ -97,9 +97,8 @@ public class SensorHubI2CConnection extends I2CConnection {
                             if (data[0] == (byte) 0xA1 || data[0] == (byte) 0xA3) {
                                 byte rpin = data[1];
 
-                                //todo: convert double.bytes to 4 instead of 8, since arduino/java are different
-                                double value = ByteBuffer.wrap(data, 2, Double.BYTES).getDouble();
-                                byte deviceSignature = data[2 + Double.BYTES];
+                                float value = ByteBuffer.wrap(data, 2, Float.BYTES).getFloat();
+                                byte deviceSignature = data[2 + Float.BYTES];
 
                                 //run a check to see if the device signature is the same as the device address, if not, ignore the data
                                 assert deviceSignature == getDeviceAddress() : "Device signature does not match device address.";
